Work around law against using a variant constructor ...
... without discriminates for an indefinite type (e.g. mytype (<>)).
It would have been nice to instantiate these directly in a way that the
uesr can manipulate, but another method is needed.

I created three overloaded "spawn_field" functions that will add a
new field to the heap.  Of the 20 types, 18 will typically use the first
spawn function, then "chains" and "settype" types, which have indefinite
sizes thus require to pass access instead, have their own spawn that also
allocates space for the chain or set.

It means the backbone of the fetchable row will likely be a string-based
hash map (based on the headers, case insensitive)