Start with documentation
This is pretty rough.  The idea has been to keep using .shtml files
for sourceforge to assemble, but to add a makefile that will generate
html files from them (just concatenate header/body/footer and sed the
.shtml to .html and save in the "generated" directory.  That's not
set up yet.

There are two pages but they are still in ".html".  That will be fixed
soon.