Implement fetch_bound
Unfortunately my attempt to use generics failed (due to the use of a
another variant record, but the end result works well.

There is a limitation to bind to the exact native time -- it will not
"convert".  It's possible, but it would make the fetch_bound as long and
complex as the result.as_* methods.