Some prep-stmt prework, add guard on "bind" procedures
It was possible to put in any positive integer for bind-by-column.
Add a check to limit it to number of columns found after execution.

The "assign" procedures are missing the same check, and worse:
We can check if the type matches at assignment, and not wait for
an "execution".  It may be necessary to move these functions
to the drivers.

At the very least, assign needs to be beefed up with index checks
as well.  Come to think of it, it should be possible to check type
at the base level.  The information should be available.