Switch error mode to logger facility
It appears the error mode code was duplicated between connections and
the logger facility.  It looks like the connections version is the
one set, but the facility version are the ones used.

This removes all traces of the connection version (including the
interface) and switches the driver to use the logger facility instead.