Fix file logging (support file name definition)
It appears that the file logging was never fully implemented.  There
was not even a default file location to log to, and no way of defining
the name of the file which was by default, blank.

Add a new function to the logger package to allow the filename to be
defined and make it log to /tmp/adabase.log in the absence of definition.