Fix query and query_select (change result type)
It appears that query was never finished.  It was set to return a base
type because that's the only way it could work as an interface (at least
as far as I can tell).

The only way I got this to compile was to change the return type to
an access of the desired statement type which required removing it off
the interface.  I couldn't use the statement because that was an
illegal copy of a limited type.

Let's see if it at least works now.