Split out column clearing to separate function
This was intended to clear data after end of data row retrieval
and discarding remaining data, but it't not possible with Ada 2005.
The fetch_bound and fetch_next routines are functions and that
would modify the Stmt which can only be done within a procedure.

For now, leave the changes in place commented out to think about
it later.  There's no net difference right now.