Rework fetch_next_set logic based on MySQL behavior
MySQL has a POLA violation about undocumented behavior.
When a stored procedure is called, the MySQL server always adds an
additional empty result set which is supposed to provide status
information about the operation.  It's not in the documentation, at all,
but I read about it in an ancient bug report.

Thus, rearrange (and duplicate) one line of code to look for this
empty result set, and if found, it will also return indication of
"no more data" even though mysql_more_results and mysql_next_result
say there is data.

Effectively this means that fetch_next_row will always stop if the
stored procedure produced an empty row set for any reason (e.g.
an insert operation).

It might be worth changing the fetch_next_rowset to have an
empty set indicator so it doesn't stop fetching prematurely
(thus rendering the rest of the stored procs inaccessible).