Take V at DECIMAL/NEW_DECIMAL
Take IV actually worked, but when MySQL converts the value, precision is
lost (float rounding errors).   Let's go back to requesting strings and
do the conversion ourselves.  We were pretty close with take III, but
the field size was set to 0 when it should been set to the value of
decimals I think.