Try to handle variable length chains
The original implementation punted if the length of the container chain
wasn't the same as the fetched data.  This is too strict; it blocks
VARBINARY and BLOB and those types of data where the length is unknown.
Instead, fill the unused blocks with zero. The application has to deal
with this.