Attempt to fix unexpected bincopy fallout
The binary copy routine for regular fetches uses the maximum length of
the blob types (e.g. 256 for tinyblob).  This isnt what we want; we want
to limit the result to the fetched size in those cases.  To fix, add a
new argument to bincopy that used fills the extract space with zeros
only if explicitly requested (as we do in PS/fetch_bound)