Try to salvage second execution
Major logic fail!  The piped execution is failing because adabase
doesn't know the type of each parameter after the statement is prepared.
They all default to BOOLEAN type.  MySQL never implemented a parameter
metadata query, so we have three choices:
1) remove the second execute feature
2) don't convert, just make all bind types a string and let mysql
convert it (not sure if that's possible but maybe)
3) Use assign ONCE for each with typed constants to define the type,
which the piped execute can then use.

So this commit defines these 20 constants (one for each type) and also
corrects a logic flaw in auto_assign.