Remove index from enumeration record
This change invalidates a bunch of enumeration converters.  The only
conversion possible now is those to strings.

Note that enumerations (and sets) are not widely supported.  SQLite and
Firebird do not support them (although they can be simulated with
foreign keys and a second table).  For those drivers, enumerations will
be treated as strings.

It's not possible to get the actual values without a second query which
I didn't realize at the time of the original design.  There will be
fallout on the testcases and documentation too.