Untested major prototype change
The fact that the statement class is limited is the source of most of my
problems I think.  Remove the limited attribute of the base statement
class so that they can be copied.  Hopefully this works towards
eliminating the major memory leaks caused by finalization occurring much
later than expected.

This will break pretty much all the test cases and require docs rework.