Interface/API change (fetch_next / iterate)
The datarows are leaking too.
Moreover, the fetch_next prototype is preventing the re-adoption of
Ada2005.

Switch to passing DataRow instead of its access.  This requires removing
the limited designation.  A new method was added, "data_exhausted" which
is a boolean that can be used like the previous version of fetch_next's
result was.

When data_exhausted is true, the datarow is the constant Empty_Datarow,
which is newly defined here.

Untested, committing to test it now.

This will require extensive fixing of testcases and documentation.