Rework column type detection
I discovered SQLite does not return one of the 5 (really 4) data types
in their metadata procedure, but rather returns the literal column text,
e.g. "varchar(45)".  The only choice is to mimic the affinity detection
the same way SQLite does it, but spread the NUMERIC affinities to other
native types.