Switch away from SQL_TRANSIENT
I have no idea why SQL_TRANSIENT is malfunctioning, but when I remove
the freemem procedure and replace it with SQL_STATIC, the bind-fetching
works (otherwise it fails after the first row is fetched).

Partial commit for testing purposes, this saves the text and blob
structures and manually frees memory later.

Aside: There appears to be no purpose to clearing column columns -- it
doesn't happen normally and finalization should take care of it.