drivers: move connection to base, abstract private_connect
Every driver had an identical call to a "private connect" so move those
up to the base driver and just call the private connect for each driver
from there.

The private connect should be abstract, but the RM doesn't allow for
private procedures to be abstract, so make it null and just make sure to
implement it for each driver.