PGSQL: disconnect the type refinement to NByteX code
It turns out that PQftype() only returns the base type Oid, so nothing
will map against constrained domains.  The only way to get it to work is
to pull every column in the database along with it's true type and store
all that information instead of just looking up the used types as we do
now.

Leave the refinement code in place but nothing calls it now.