Tweak bad_select testcase so it behaves identically in all drivers
Before autocommit mode off (the default) so a transaction is in
progress.  PostgreSQL will not allow a transaction to continue after
a failed command, even a failed select command.  The options are
A) turn off autocommit, B) rollback everything C) implement auto-save
point in AdaBase for PostgreSQL.  While (C) is interesting, it doesn't
make a lot of sense since AdaBase is not interactive, so we go with
option (A).