Narrow down permission of risky binding
Allowing unsigned variables to bind to unsigned values is always risk
(there's no guarantee it's constrainted at db level), so only allow this
for comparable sizes (e.g. allow NByte2 to find to Byte2, but don't
allow NByte3 to bind with Byte2).
Also change message when sizing is the reason for rejection.