Create the ability to spawn null fields
This is probably needed by the other two drivers.
When the data returned is "null", then many string conversions are going
to fail.  We need to set the values to same default along with "null"
attribute instead, so creating a fourth "spawn_field" function is
necessary.  I'll have to document this later.  It also merits a new
testcase.