I keep learning new things about postgresql
For example: If you DEALLOCATE a prepared statement during an ongoing
transaction, the entire transaction is invalidated.  I don't see the
logic.

The workaround is to allow deallocation on the spot only when in
autocommit mode.  If in the standard mode (autocommit off) then the
deallocation is deferred until the next rollback or commit.  Closing the
connection would deallocate anything left.