Recover pgresult earlier
Make sure that when data is "exhausted" we recover the pgresult
resources right then, and also (and more importantly), when the user
specifies to discard the rest of the data.

This was not the source of the small memory leak, however.  The
finalization routine eventually collected it when the statement
finalized.