Add new type "bits" which is an array of bits
While here, remove endian handling.  The second chain link is always
bits 8-15, the third is 16-23, etc.

The bit type index starts with zero, which is typical for arrays of bits
but different than chains and other array types.

As a result, some functions could be made into generic functions and
this was done.