Fix character setting logic flaw
We need to determine if the character set is UTF8 encoded by actually
querying the database after we set it.  This is because a blank string
can still result in UTF8 if the database is UTF8 by default.

Caught by PostgreSQL but MySQL had the same flaw.  SQLite is always UTF8
so it wasn't affected.