Seems GNAT has a bug, this is large, inconvenient workaround
I assumed the variant record GEOMETRY to the variant record for holding
all the different data types.  It compiles fine and appears legal, but
in 2 locations, it causes an "Illegal Instruction" error (the variant
field record and the variant bind record).  I wasn't successful in
making a small test case.

I believe this is a nasty GNAT bug.  However, I don't have time to track
this down, so I'm changing the plan after losing one day on this:
1) change the variant entry from GEOMETRY to TEXTUAL
2) Store the Well_Known_Binary representation in this field.  It seems
both MySQL and PostgreSQL return this
3) generate GEOMETRY types on the fly with "as_geometry" and bind
functions
4) I'll need to create a GEOMETRY => WKB converter, which will be quite
painful for the IEEE 754 form I think.  This is needed for parameter
assignments and inserting data.

This is untested.  I'm crossing my fingers the ILL goes away.