FINALLY!
After hours and hours, I think I squashed the elusive multipoint in geo
collection inside geo collection bug.  I had to implement the ring count
again and fix a few bugs along the way (e.g. size of collection is not
the same as size of single* and multi*.  the former counts groups and
the later counts units, that's a big flaw).  there's another couple of
bug fixes in here too.